+++ src/greensenti/__main__.py
@@ -25,11 +25,8 @@ def main():
         },
         "raster": {"apply-mask": raster.apply_mask, "transform-image": raster.transform_image},
         "download": {
-            # Fire has an issue returning pandas.DataFrames, those functions
-            # must return a str instead
-            # Reference: https://github.com/google/python-fire/issues/274
-            "by-title": lambda *args, **kwargs: str(dhus.download_by_title(*args, **kwargs)),
-            "by-geometry": lambda *args, **kwargs: str(dhus.download_by_geometry(*args, **kwargs)),
+            "by-title": dhus.download_by_title,
+            "by-geometry": dhus.download_by_geometry,
         },
     }
     fire.Fire(cli_map)